Added guaranteed/immedate message types
