+++ Example/WatchExample Extension/ExtensionDelegate.swift
@@ -71,7 +71,7 @@ private extension ExtensionDelegate {
     }
     
     private func setupMessageReceivedObservers() {
-        Communicator.shared.messageReceivedObservers.add { message in
+        Communicator.shared.immediateMessageReceivedObservers.add { message in
             print("Received message: \(message)")
             message.replyHandler?(["Replied!" : "Message"])
             self.endWatchConnectivityBackgroundTaskIfNecessary()
